refactored some of the autoencoder experiment code for clean up.
