+++ experiments/autoencoders/autoencoders_besthps.py
@@ -4,147 +4,23 @@ import os
 import sys
 import time
 from functools import partial
-from typing import Callable, List, NamedTuple, Union
 
 import numpy as np
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
-import torch.optim as optim
-import torchvision
-import torchvision.transforms as transforms
+
 from ray import air, tune
 from ray.air import session
 from torch.utils.tensorboard import SummaryWriter
 
-from datasets import CURVESDataset, FACESDataset
 
 sys.path.insert(0, os.path.expanduser("~/experiments"))
 
 from kradagrad.utils import get_optimizer
 
-
-class DenseNet(nn.Module):
-    """Implements a fully connected encoder-decoder network.
-    """
-    def __init__(self, encoder_widths, decoder_widths, act_fn=nn.ReLU(), out_fn=None):
-        super(DenseNet, self).__init__()
-
-        assert (
-            encoder_widths[-1] == decoder_widths[0]
-        ), "encoder output and decoder input dims must match"
-
-        enc_layers = {}
-        for k in range(len(encoder_widths) - 1):
-            enc_layers[f"enc_layer_{k}"] = nn.Linear(
-                encoder_widths[k], encoder_widths[k + 1]
-            )
-        self.enc_layers = nn.ModuleDict(enc_layers)
-
-        dec_layers = {}
-        for k in range(len(decoder_widths) - 1):
-            dec_layers[f"dec_layer_{k}"] = nn.Linear(
-                decoder_widths[k], decoder_widths[k + 1]
-            )
-        self.dec_layers = nn.ModuleDict(dec_layers)
-
-        self.act_fn = act_fn
-        self.out_fn = out_fn
-
-    def forward(self, x):
-        for k in range(len(self.enc_layers)):
-            x = self.enc_layers[f"enc_layer_{k}"](x)
-            x = x if k == len(self.enc_layers) - 1 else self.act_fn(x)
-
-        for k in range(len(self.dec_layers)):
-            x = self.dec_layers[f"dec_layer_{k}"](x)
-            if k == len(self.dec_layers) - 1:
-                x = x if self.out_fn is None else self.out_fn(x)
-            else:
-                x = self.act_fn(x)
-        return x
-
-
-class MNISTConfig(NamedTuple):
-    encoder_widths = [784, 1000, 500, 250, 30]
-    decoder_widths = [30, 250, 500, 1000, 784]
-    act_fn = nn.ReLU()
-    out_fn = None
-    loss_fn = nn.CrossEntropyLoss()
-
-
-class FACESConfig(NamedTuple):
-    encoder_widths = [625, 2000, 1000, 500, 30]
-    decoder_widths = [30, 500, 1000, 2000, 625]
-    act_fn = nn.ReLU()
-    out_fn = None
-    loss_fn = nn.MSELoss()
-
-
-class CURVESConfig(NamedTuple):
-    encoder_widths = [784, 400, 200, 100, 50, 25, 6]
-    decoder_widths = [6, 25, 50, 100, 200, 400, 784]
-    act_fn = nn.ReLU()
-    out_fn = None
-    loss_fn = nn.CrossEntropyLoss()
-
-
-def get_task_cfg(dataset):
-    if dataset == "mnist":
-        return MNISTConfig()
-    elif dataset == "faces":
-        return FACESConfig()
-    elif dataset == "curves":
-        return CURVESConfig()
-    else:
-        raise ValueError(f"Dataset {dataset} is not supported")
-
-
-def get_dataloaders(dataset, batch_size=100, root=None):
-    if dataset == "mnist":
-        train_set = torchvision.datasets.MNIST(
-            root=root or "./data",
-            train=True,
-            download=True,
-            transform=transforms.ToTensor(),
-        )
-        test_set = torchvision.datasets.MNIST(
-            root=root or "./data",
-            train=False,
-            download=True,
-            transform=transforms.ToTensor(),
-        )
-
-    elif dataset == "faces":
-        train_set = FACESDataset(
-            root=root or "./data",
-            transform=transforms.Compose(
-                [transforms.ToTensor(), transforms.Normalize(mean=0.0, std=1.0)]
-            ),
-        )
-        train_set, test_set = torch.utils.data.random_split(train_set, [103500, 62100])
-
-    elif dataset == "curves":
-        train_set = CURVESDataset(
-            root=root or "./data",
-            train=True,
-        )
-
-        test_set = CURVESDataset(
-            root=root or "./data",
-            train=False,
-        )
-    else:
-        raise ValueError(f"Dataset {dataset} is not supported")
-
-    train_loader = torch.utils.data.DataLoader(
-        train_set, batch_size=batch_size, shuffle=True, num_workers=2
-    )
-    test_loader = torch.utils.data.DataLoader(
-        test_set, batch_size=batch_size, shuffle=False, num_workers=2
-    )
-
-    return train_loader, test_loader
+from .densenet import DenseNet
+from .configs import get_task_cfg
+from .dataloaders import get_dataloaders
 
 
 def main(tune_cfg, args):
